Filter out inline functions
Instead of using "meta.function" as the selector for function regions in a
file, let's instead use "meta.function - meta.function.inline" to filter
out any inline functions. This might be better suited to hide behind a
preference, but I'm putting it inline for now as an example of what we can
do with the selector engine.